Bug 1928723 - Callback kind refactor, r=bgruber
Updated/refactored `config.toml` handling to also configure callback/trait
interface methods.  Right now the valid variants for these are `Async`
and `FireAndForget`, but I hope to enable support for sync soon.

Replaced `CppCallbackInterfaceMethod::async_data` with
`CppCallbackInterfaceMethod::kind`.  Again the two variants are
`Async`/`FireAndForget`, enabling `Sync` is next.

Updated the ApplicationErrorReporter callbacks to be `FireAndForget`
instead of `AsyncWrapped`.  I believe `FireAndForget` is correct and the
old config was only working because we weren't checking it for callback
methods.